move event_loop to separate file
