+++ starknet_py/conftest.py
@@ -1,5 +1,6 @@
 # This is needed for importing fixtures from `fixtures` directory
 pytest_plugins = [
+    "starknet_py.tests.e2e.fixtures.event_loop",
     "starknet_py.tests.e2e.fixtures.clients",
     "starknet_py.tests.e2e.fixtures.account_clients",
     "starknet_py.tests.e2e.fixtures.contracts",
