+++ starknet_py/tests/e2e/fixtures/event_loop.py
@@ -1,0 +1,11 @@
+# This fixture was added to enable using async fixtures
+import asyncio
+import pytest
+
+
+@pytest.fixture(scope="module")
+def event_loop():
+    policy = asyncio.get_event_loop_policy()
+    loop = policy.new_event_loop()
+    yield loop
+    loop.close()
